requirements.txt
Flask==2.2.3
python-telegram-bot==20.1
pymongo==4.4.0
python-dotenv==1.0.0
requests==2.28.1
